requirements.txt
streamlit
requests
python-dotenv
langchain
langchain-community
langsmith
langchain-ollama
openai
langchain-openai